requirements.txt
fastapi==0.123.10
uvicorn==0.38.0
python-dotenv==1.2.1
pymongo==4.15.5
passlib[bcrypt]==1.7.4
python-jose==3.5.0
python-multipart==0.0.20
pydantic==2.12.5
pydantic-settings==2.12.0
email-validator==2.3.0
dnspython==2.8.0
certifi==2025.11.12
